+++ UVRPN_Unity/Assets/Scripts/Core/VRPN_Analog.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using UVRPN.Events;
 
 namespace UVRPN.Core
 {