+++ UVRPN_Unity/Assets/Scripts/Utility/AnalogEvent.cs
@@ -1,7 +1,0 @@
-﻿using System;
-using UnityEngine;
-using UnityEngine.Events;
-
-[Serializable]
-public class AnalogEvent : UnityEvent<Vector2> {
-}